Restrict memory and CPU for our support services
The intent of this is to prevent support services from consuming
system resources to a point that it interferes with the main service
running.

Main services should _also_ be constrained, such that they cannot
interfere with support services. This dual set of constraints should
mean that even under extreme load scenarios all services will continue
to operate in a manner that allows us to diagnose and potentially resolve
the underlying conditions causing elevated load.