Make teak-log-collector a soft dependency for teak-init
If the log collector fails we should still have logs, just stored on
the server. We should also be able to detect this situation -- systemd
logs stop showing up for a server still in the running state.

Given this, I think "we're not phoning home logs" shouldn't be an issue
that immediately takes down the entire server. There might be value in
detecting "and we're never going to phone home logs again" and killing
the server in that case though.